requirements.txt
colorlog==3.1.4
humanfriendly==4.12
termcolor==1.1.0
progressbar2==3.38.0